add initset support, cleanup file structure
